Allow groups to be private (#810)
The permissions are not sophisticated. A group can be either public or private. From a user's perspective, private groups don't exist; users can't create private groups, change the privacy setting of groups, or see, modify, or be aware of the existence of a group that's private, even if they're an organizer of it. Private groups do show up in the admin; admins can create private groups and can see and change everything about every group, including privacy settings.

Fixes: #776