Migrate to reCAPTCHA v3
The threshold is currently set at 85%; we can adjust it up or down later if need be.

This involves creating a new site in the reCAPTCHA admin console; I currently own it.

The site key is in a separate setting from the v2 one, which I've already created in production, so we can switch back and forth with just code deployments if need be.

By default, reCAPTCHA is disabled in local development, but you can enable it by signing up for a site on localhost and adding your keys to your .env file.

I don't know how much this will help, but it may be worth a try.